Data.DBVar

Description

Synopsis

DBVar represents a mutable variable whose value is kept in memory, but which is written to the hard drive on every update. This provides a convenient interface for persisting values across program runs. For efficient updates, delta encodings are used, see Data.Delta.

Store represent a storage facility to which the DBVar is written.

DBVar

data DBVar (m :: Type -> Type) delta

A DBVar m delta is a mutable reference to a Haskell value of type a. The type delta is a delta encoding for this value type a, that is we have a ~ Base delta.

The Haskell value is cached in memory, in weak head normal form (WHNF). However, whenever the value is updated, a copy of will be written to persistent storage like a file or database on the hard disk; any particular storage is specified by the Store type. For efficient updates, the delta encoding delta is used in the update.

Concurrency: DBVar fully supports concurrent reads and updates.

Updates are atomic and will block other updates.

Reads will not be blocked during an update (except for a small moment where the new value atomically replaces the old one).

readDBVar :: (Delta da, a ~ Base da) => DBVar m da -> m a

Read the current value of the DBVar.

updateDBVar :: (Delta da, Monad m) => DBVar m da -> da -> m ()

Update the value of the DBVar using a delta encoding.

The new value will be evaluated to weak head normal form.

modifyDBVar :: (Delta da, Monad m, a ~ Base da) => DBVar m da -> (a -> (da, b)) -> m b

Modify the value in a DBVar.

The new value will be evaluated to weak head normal form.

modifyDBMaybe :: (Delta da, Monad m, a ~ Base da) => DBVar m da -> forall b. (a -> (Maybe da, b)) -> m b

Maybe modify the value in a DBVar

If updated, the new value will be evaluated to weak head normal form.

initDBVar

Arguments

:: (MonadSTM m, MonadThrow m, MonadEvaluate m, MonadMask m, Delta da, a ~ Base da)

=> UpdateStore m da
Store for writing.

-> a
Initial value.

-> m (DBVar m da)

Initialize a new DBVar for a given Store.

loadDBVar

Arguments

:: (MonadSTM m, MonadThrow m, MonadEvaluate m, MonadMask m, Delta da)

=> UpdateStore m da
Store for writing and for reading the initial value.

-> m (DBVar m da)

Create a DBVar by loading its value from an existing Store. Throws an exception if the value cannot be loaded.