Data.Delta.Update

Description

Synopsis

Update represents a computation which produces a delta and a result.

Note: This module is preliminary.

Update

Type

data Update da r

A computation which inspects a value a ~ Base da and produces a delta da and a result of type r.

Similar to the State computation, but involving Delta types.

View

runUpdate :: a ~ Base da => Update da r -> a -> (Maybe da, r)

Run the Update computation.

applyUpdate :: (Delta da, a ~ Base da) => Update da r -> a -> (a, r)

Semantics.

onDBVar :: (Monad m, Delta da) => DBVar m da -> Update da r -> m r

Apply an Update to a DBVar.

Combinators

nop :: Update da ()

No operation.

Use the Functor instance, specifically (<$) to add results other than ().

update :: a ~ Base da => (a -> da) -> Update da ()

Compute a delta.

updateWithResult :: a ~ Base da => (a -> (da, r)) -> Update da r

Compute a delta with result.

Helpers

updateWithError :: a ~ Base da => (a -> Either e da) -> Update da (Either e ())

Computer a delta or fail.

updateWithResultAndError :: a ~ Base da => (a -> Either e (da, r)) -> Update da (Either e r)

Compute a delta with result or fail.

updateMany :: Update da r -> Update [da] r

Lift an update for a single delta to a list of deltas.

updateField

Arguments

:: (a ~ Base da, b ~ Base db)

=> (b -> a)
View field.

-> (da -> db)
Lift delta to

-> Update da r

-> Update db r

Helper function for lifting the Update from a record field to the record.

Example:

data Pair a b = Pair a b
first :: Pair a b -> a

data DeltaPair da db
    = UpdateFirst da
    | UpdateSecond db

updateField first UpdateFirst
    :: (a -> Update da r)
    -> (Pair a b -> Update (DeltaPair da db) r)