Test.Data.Store

Description

This module provides utilities for testing Store implementations.

prop_StoreUpdate is a general property test that tests the laws for updateS.

genChain generates random sequences of deltas.

StoreUnitTest provides a monadic DSL for writing example test cases for Store.

Store laws

type GenDelta da = Base da -> Gen da

Given a value, generate a random delta that applies to this value.

prop_StoreUpdate

Arguments

:: forall m da (qa :: Type -> Type). (Monad m, Delta da, Eq (Base da), Show da, Show (Base da))

=> (forall b. m b -> IO b)
Function to embed the monad in IO

-> m (Store m qa da)
Creation for Store that is to be tested.

-> Gen (Base da)
Generator for the initial value.

-> GenDelta da
Generator for deltas.

-> Property

Test whether the law on updateS is satisfied.

Subsumes test for the law on writeS / loadS.

Generators

data Chain da

Chain of deltas and the results of their application.

The delta that is applied *last* appears in the list *first*.

Constructors

Chain [(Base da, da)] (Base da)

genChain :: Delta da => Gen (Base da) -> GenDelta da -> Gen (Chain da)

Randomly generate a chain of deltas.

shrinkChain :: Chain da -> [Chain da]

Shrink a chain of deltas.

Unit test DSL for developing a Store

data StoreUnitTest (m :: Type -> Type) (qa :: Type -> Type) da r

A monadic DSL to unit test a Store.

unitTestStore :: forall m da (qa :: Type -> Type) a. (Monad m, Eq (Base da), Show (Base da), Show da) => Base da -> Store m qa da -> StoreUnitTest m qa da a -> m Property

Run a unit test for a Store.

applyS :: forall (m :: Type -> Type) da (qa :: Type -> Type). (Monad m, Delta da) => da -> StoreUnitTest m qa da ()

Apply a delta to the current value.

checkLaw :: forall (m :: Type -> Type) da (qa :: Type -> Type). (Monad m, Eq (Base da), Show (Base da), Show da) => StoreUnitTest m qa da ()

Check the store laws.

reset :: forall (m :: Type -> Type) (qa :: Type -> Type) da. Monad m => StoreUnitTest m qa da ()

Reset the store state to the initial value.

context :: forall (m :: Type -> Type) (qa :: Type -> Type) da x. Monad m => (Property -> Property) -> StoreUnitTest m qa da x -> StoreUnitTest m qa da x

Add a context to test.

observe :: forall (m :: Type -> Type) da (qa :: Type -> Type). Monad m => (Base da -> Property) -> StoreUnitTest m qa da ()

Observe a property on the current value of the store.

ignore :: forall (m :: Type -> Type) (qa :: Type -> Type) da x. Monad m => StoreUnitTest m qa da x -> StoreUnitTest m qa da x

Ignore the properties of a sub-test.